Rearrange code and more vg work
