--- etcd/proto/README
@@ -1,1 +1,0 @@
-Need unit tests!
